ml-service/ml-requirements.txt
# ML Service Dependencies
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# ML и AI библиотеки
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
transformers>=4.30.0
torch>=2.0.0
openai>=1.0.0

# Text processing
langdetect>=1.0.9
textblob>=0.17.1

# HTTP клиент
httpx>=0.24.0
